CDHSH-26 Refactored Filter Header
