+++ services/frontend/src/scenes/People/PeopleLayout.js
@@ -1,7 +1,6 @@
 import React from "react";
 import "./People.scss";
 import {FilterHeader, ProfileCard} from "components/";
-import classNames from "classnames";
 
 const sortProfiles = (profiles) => profiles.sort((a, b) => a.name.localeCompare(b.name));
 
@@ -14,28 +13,6 @@ const PeopleLayout = ({profiles, activeFilter, onFilterClick}) => (
         />
         <PeopleList people={sortProfiles(profiles)} />
     </div>
-);
-
-const PeopleHeader = ({activeFilter, onFilterClick}) => (
-    <div className="people-header">
-        <FilterButton
-            label="All"
-            isActive={activeFilter}
-            onClick={onFilterClick}
-        />
-    </div>
-);
-
-const FilterButton = ({label, isActive = false, onClick}) => (
-    <button
-        className={classNames(
-            "filter-button",
-            {"filter-button--active": isActive}
-        )}
-        onClick={onClick}
-    >
-        {label}
-    </button>
 );
 
 const PeopleList = ({people}) => {
