+++ services/frontend/src/scenes/Projects/ProjectsLayout.js
@@ -9,47 +9,13 @@ const ProjectsLayout = ({projects, activeFilter, onFilterClick}) => (
         <FilterHeader
             activeFilter={activeFilter}
             onFilterClick={onFilterClick}
-            labels={[Project.FILTER_ALL, Project.FILTER_ACTIVE, Project.FILTER_ALL]}
+            labels={[Project.FILTER_ALL, Project.FILTER_ACTIVE, Project.FILTER_INACTIVE]}
         />
 
         <ProjectsList
             projects={projects}
         />
     </div>
-);
-
-const ProjectsHeader = ({activeFilter, onFilterClick}) => (
-    <div className="projects-header">
-        <FilterButton
-            label="All"
-            isActive={activeFilter === Project.FILTER_ALL}
-            onClick={onFilterClick(Project.FILTER_ALL)}
-        />
-
-        <FilterButton
-            label="Active"
-            isActive={activeFilter === Project.FILTER_ACTIVE}
-            onClick={onFilterClick(Project.FILTER_ACTIVE)}
-        />
-
-        <FilterButton
-            label="Inactive"
-            isActive={activeFilter === Project.FILTER_INACTIVE}
-            onClick={onFilterClick(Project.FILTER_INACTIVE)}
-        />
-    </div>
-);
-
-const FilterButton = ({label, isActive = false, onClick}) => (
-    <button
-        className={classNames(
-            "filter-button",
-            {"filter-button--active": isActive}
-        )}
-        onClick={onClick}
-    >
-        {label}
-    </button>
 );
 
 const ProjectsList = ({projects}) => {
